test(extension): cover sidecar schema declarations
